Moves 'getUsers' out of index.js file
