+++ src/controllers/index.js
@@ -1,17 +1,5 @@
-const { findUsers } = require("../model")
-
-const getUsers = async (req, res, next) => {
-    try {
-        const users = await findUsers()
-        res.json(users)
-    } catch (error) {
-        console.log(error)
-        next(error)
-    }
-}
-
 module.exports = {
-    getUsers,
+    getUsers: require("./get-users"),
     login: require("./login"),
     addUser: require("./add-user"),
     updateUser: require("./update-user"),
